+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     01. :boom: Methods of `dev.orne.config.PreferencesConfig` throw `ConfigException`
     01. :boom: Methods of `dev.orne.config.PropertiesConfig` throw `ConfigException`
     01. :boom: Methods of `dev.orne.config.SystemConfig` throw `ConfigException`
+01. :gift: Added `dev.orne.config.DelegatedConfig`
 
 # :package: 0.1.0